goal input component with props
