--- App.js
@@ -1,90 +1,0 @@
-import React, { useState } from 'react'
-import { StyleSheet, Text, View, Button, TextInput, ScrollView, FlatList } from 'react-native'
-
-
-// Application Screen || Define Exports
-// =================================================================================================
-// =================================================================================================
-export default function App() {
-  const [enteredGoal, setEnteredGoal] = useState("")
-  const [goals, setGoals] = useState([])
-  
-  const goalInputHandler = (enteredText) => {
-    setEnteredGoal(enteredText)
-  }
-
-  const addGoalHandler = () => {
-    setGoals(currentGoals => [
-      ...currentGoals,
-      { text: enteredGoal, key: Math.random().toString() }
-    ])
-  }
-
-  return (
-    <View style={styles.appContainer}>
-      <View style={styles.inputContainer}>
-        <TextInput
-          style={styles.textInput}
-          placeholder="Your Fitness Goal"
-          onChangeText={goalInputHandler}
-        />
-        <Button
-          title="Add Goal"
-          onPress={addGoalHandler}
-        />
-      </View>
-      <View style={styles.goalsContainer}>
-        <FlatList
-          data={goals}
-          renderItem={itemData => {
-            return (
-              <View style={styles.goalItem}>
-                <Text style={styles.goalText}>{itemData.item.text}</Text>
-              </View>
-            )
-          }}
-          // keyExtractor={(item, index) => {
-          //   return item.id
-          // }}
-        />
-      </View>
-    </View>
-  )
-}
-
-const styles = StyleSheet.create({
-  appContainer: {
-    flex: 1, // This will make the app container take up the entire screen
-    marginTop: 10,
-    paddingTop: 50,
-    paddingHorizontal: 16,
-  },
-  inputContainer: {
-    flex: 1, // This will make the input container take up 1/4 of the screen
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 24,
-    borderBottomWidth: 1,
-    borderBottomColor: '#cccccc',
-  },
-  textInput: {
-    borderWidth: 1,
-    borderColor: "#cccccc",
-    width: "70%",
-    marginRight: 8,
-    padding: 8,
-  },
-  goalsContainer: {
-    flex: 5, // This will make the goals container take up 5/6 of the screen
-  },
-  goalItem: {
-    margin: 2,
-    padding: 10,
-    borderRadius: 6,
-    backgroundColor: "aliceblue",
-  },
-  goalText: {
-    color: "black",
-  }
-})
